requirements.txt
Django==3.0.2
psycopg2==2.8.4
djangorestframework==3.11.0
django-cors-headers